backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
supabase==2.9.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
anthropic==0.39.0
openai==1.54.0
apscheduler==3.10.4
redis==5.1.1
slowapi==0.1.9
sendgrid==6.11.0
python-dotenv==1.0.1
